cleanup outdated code
